+++ packages/@expo/cli/add-module.js
@@ -1,8 +1,0 @@
-const { loadModule } = require('@expo/require-utils');
-
-/**
- * @deprecated Will be removed in a future release.
- */
-module.exports = function importESM(name) {
-  return loadModule(name);
-};